Migrasi database MySQL ke MongoDB (mongo-go-driver)
- Semua repository ditulis ulang: guests, conversations, tickets,
  hotel_information, recommendations, staff, ticket_assignments
- ID integer atomik via collection counters (format TKT-xxxxxx tetap)
- Index + seed idempotent di repository.Migrate (ganti migrations/*.sql)
- Config: MONGO_URI + DB_NAME (ganti DB_DSN MySQL)
- Test pakai database terisolasi per test (batiqa_test_<nama>)
- Fix assign tiket: staff_id body dipakai untuk admin assign-atas-nama
- docker-compose, run.bat, .env.example, README sesuai MongoDB
- E2E lolos vs MongoDB lokal: login, chat->tiket, info hotel dari DB,
  rekomendasi berbudget, transisi status, prioritas, assign, stats

+++ internal/handler/stats.go
@@ -16,12 +16,12 @@ func NewStatsHandler(repo *repository.TicketRepository) *StatsHandler {
 }
 
 type StatsResponse struct {
-	TotalOpen     int `json:"total_open"`
-	HighPriority  int `json:"high_priority"`
-	Housekeeping  int `json:"housekeeping"`
-	Engineering   int `json:"engineering"`
-	ResolvedToday int `json:"resolved_today"`
-	TotalTickets  int `json:"total_tickets"`
+	TotalOpen     int64 `json:"total_open"`
+	HighPriority  int64 `json:"high_priority"`
+	Housekeeping  int64 `json:"housekeeping"`
+	Engineering   int64 `json:"engineering"`
+	ResolvedToday int64 `json:"resolved_today"`
+	TotalTickets  int64 `json:"total_tickets"`
 }
 
 func (h *StatsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
